Styling improvements

+++ MyBooks/Models/Book/BookDetailVM.cs
@@ -8,7 +8,7 @@ public class BookDetailVM
     public string ISBN { get; set; }
     public string Description { get; set; }
     public string ThumbnailUrl { get; set; }
-    public double Rating { get; set; }
+    public int Rating { get; set; }
 
     public string? Library { get; set; }
     public List<ReviewVM>? Reviews { get; set; }
@@ -18,7 +18,7 @@ public class ReviewVM
 {
     public Guid Id { get; set; }
     public string Content { get; set; }
-    public double Rating { get; set; }
+    public int Rating { get; set; }
     public DateTime CreatedAt { get; set; }
     public string User { get; set; }
 }
